requirements.txt
streamlit==1.33.0
networkx==3.3
Pyvis==0.3.2
PyPDF2==3.0.1
termcolor==2.4.0
python-dotenv
langchain
langchain-text-splitters   
langchain-community
pinecone-client 
pinecone-text 
langchain-core
pinecone-notebooks 
langchain-huggingface 
sentence-transformers 
tf-keras
langchain-openai
nltk
neo4j
anthropic
openai
pycrypto
asyncio